scMD framework

This function performs signature generation and ensemble deconvolution using the scMD algorithm.

scMD(
  bulk,
  bulk_type = "450k_or_850k",
  ncluster = 5,
  dmet_list = NULL,
  nmrk = 100,
  enableFileSaving = FALSE,
  output_path = NULL
)

Arguments

bulk

A matrix of bulk data, with CpGs by sample.

bulk_type

A character string indicating the technical platform for the bulk data. Available options include "450k_or_850k" or "WGBS",.

ncluster

An integer specifying the number of cores for parallel computation. Default is 5.

dmet_list

A character vector specifying the desired deconvolution methods to be used. If NULL (the default), the following methods will be included: c("CIBERSORT", "EPIC", "FARDEEP", "DCQ", "ICeDT", "Houseman", "NNLS", "RPC").

nmrk

The number of markers per cell type used in deconvolution. Default is 100.

enableFileSaving

Enable Saving of Intermediate Output (Optional) A boolean flag that controls the saving of intermediate outputs as separate files. When set to TRUE, intermediate outputs of the analysis will be saved to files. If not explicitly set, this parameter defaults to FALSE, meaning that intermediate outputs will not be saved by default.

output_path

A character specifying the path for saving intermediate results. Default is NULL.

Value

If scMD converges, return:

scMD_p: matrix, scMD estiamted CTS fractions.

phat_all: list, estimated cell-type fractions.

Examples

data("Guintivano")
Est.prop.scMD = scMD(bulk = Guintivano_bulk_sub, bulk_type = "450k_or_850k",output_path = "~/Desktop/")